wip update tests (moved fixtures in separate modules)
